requirements.txt
streamlit==1.39.0
requests==2.32.3
folium==0.17.0
python-dotenv==1.0.1
pytz==2024.2
sqlite3